pyproject.toml
[project]
name = "darwincore"
version = "0.1.0"
description = "Generator for extended/event-based DarwinCore-Archive (DwC-A) files."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "dict2xml>=1.7.6",
    "jinja2>=3.1.6",
    "pytz>=2025.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "pytest>=8.3.5",
    "ruff>=0.11.0",
]

[project.scripts]
dwca-generator-cli = "darwincore.dwca_generator_cli:main"
dwca-generator-main = "darwincore.dwca_generator_main:main"
publish-to-yamr = "darwincore.metadataopendata.publish_to_yamr:main"
put-metadatafile-to-yamr = "darwincore.metadataopendata.put-metadatafile-to-yamr:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 90
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I001",
    "RUF"
]

